Moved authorization classes to new folders in the shared project.
