+++ FribergFastigheter.Shared/FribergFastigheter.Shared.csproj
@@ -10,4 +10,8 @@
     <PackageReference Include="Microsoft.AspNetCore.Authorization" Version="8.0.2" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Services\AuthorizationHandlers\Broker\" />
+  </ItemGroup>
+
 </Project>
