+++ FribergFastigheter.Shared/Services/AuthorizationHandlers/Housing/Data/IDeleteHousingAuthorizationData.cs
@@ -1,11 +1,11 @@
-﻿namespace FribergFastigheter.Shared.Services.AuthorizationHandlers.Data.Housing
+﻿namespace FribergFastigheter.Shared.Services.AuthorizationHandlers.Housing.Data
 {
     /// <summary>
-    /// Interface designed to be used with the <see cref="ManageHousingAuthorizationHandler"/> to enable authorization for creating housing image objects.
+    /// Interface designed to be used with the <see cref="ManageHousingAuthorizationHandler"/> to enable authorization for deleting housing objects.
     /// </summary>
     /// <!-- Author: Jimmie -->
     /// <!-- Co Authors: -->
-    public interface ICreateHousingImageAuthorizationData
+    public interface IDeleteHousingAuthorizationData
     {
         /// <summary>
         /// The existing housing broker firm ID.
